+++ doc/conf.py
@@ -15,13 +15,11 @@ import sys
 
 sys.path.insert(0, os.path.abspath("../"))
 
-
 # -- Project information -----------------------------------------------------
 
 project = "Exasol Toolbox"
-copyright = "2022, Exasol"
+copyright = "2022, Exasol"  # pylint: disable=redefined-builtin
 author = "Exasol"
-
 
 # -- General configuration ---------------------------------------------------
 
@@ -60,7 +58,6 @@ templates_path = ["_templates"]
 # This pattern also affects html_static_path and html_extra_path.
 exclude_patterns = ["_build", "Thumbs.db", ".DS_Store", ".build-docu"]
 
-
 # -- Options for HTML output -------------------------------------------------
 
 # The theme to use for HTML and HTML Help pages.  See the documentation for
@@ -72,7 +69,7 @@ html_theme = "furo"
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
 html_static_path = ["_static"]
-html_title = f"Toolbox"
+html_title = "Toolbox"
 html_theme_options = {
     "light_logo": "light-exasol-logo.svg",
     "dark_logo": "dark-exasol-logo.svg",